Simple MS-DOS emulator.
In an effort to get Racter running where I can pipe I/O to it, I realize it
might be easier to fake enough MS-DOS to get it running.  It appears to use
nothing more than standard MS-DOS calls (from what I can gather) so if I'm
lucky, it won't be more than open(), read(), write() and close() (or rather,
the MS-DOS equivalents of those.

To get that, I have a 32-bit Linux system, which has the vm86() system call.
There isn't much about that on the Intarwebs, but given that I now have a
working example of using vm86(), I think I can proceed from here.

And to think, this is *easier* than patching DOSemu or DOSBox to accept
input via a pipe.