Get Eliza to work with Racter.
This is a specialized Eliza that knows how to "talk" to Racter.  There's no
buffering on input or output, and this Eliza knows to start talking to
Racter with its name.  Also, this Eliza keeps track of how many words have
passed and after 50,000+ words, attempts to quit Racter.  This may or may
not work due to a bug in Racter that it always doesn't quite understand when
we want to quit.

Hopefully, this mess will work the first time.

I doubt it.