Last ditch effort to get this working.
I'm convinced there's a race condition or deadlock somewhere.  Sometimes
Racter will just up and quit for no reasons I can at this time determine.
Othertimes the programs will just stop talking, where each side is
apparently waiting for the other side to talk.  This is a complete bodge
anyway so it's not surprising it has ...  issues.

I'm stopping here, at 15,911 words.  I could continue for another hour
or so trying to up the wordcount to 50,000 but let's face it, it's pretty
bad output.  So I'm calling this a wash.