Rename the Eliza 'script'.
Cleaning up the repository for posting, and I don't really need anything in
this directory except for the script.  It appears that I'm losing the actual
history but there isn't much history anyway, so ...  whatever.

Anyway, this is apparently a very close (if not the actual) version of Eliza
as originally written.  I've added verbiage describing how to interpret the
script and maybe one of these days I'll get around to writing code that does
so.